refactor(layout): extract column bottom note sorting
Move bottom-note detection and column-run sorting into a dedicated helper. This keeps the multi-column reordering logic focused on column detection and buffering while preserving the late-note ordering behavior.
